Remove the 'final' qualifier restriction for Resource classes.
At this point we put all responsability on the user to not abuse the
resource classes. Since the main concept behind the Resource classes is
namespace bassed, we don't need to care if the class is final or not.
This also alow to mock resources on the consumer side, wich is a good
thing for testing.